attempt to consolidate update logic
Running into weird windows/node/idk issues
